+++ projects/openttd_vs100.vcxproj
@@ -1211,11 +1211,10 @@
     <ClInclude Include="..\src\network\core\os_abstraction.h" />
     <ClCompile Include="..\src\network\core\packet.cpp" />
     <ClInclude Include="..\src\network\core\packet.h" />
-    <ClCompile Include="..\src\network\core\tcp.cpp" />
-    <ClInclude Include="..\src\network\core\tcp.h" />
     <ClCompile Include="..\src\network\core\tcp_admin.cpp" />
     <ClInclude Include="..\src\network\core\tcp_admin.h" />
-    <ClCompile Include="..\src\network\core\tcp_connect.cpp" />
+    <ClCompile Include="..\src\network\core\tcp_connecter.cpp" />
+    <ClInclude Include="..\src\network\core\tcp_connecter.h" />
     <ClCompile Include="..\src\network\core\tcp_content.cpp" />
     <ClInclude Include="..\src\network\core\tcp_content.h" />
     <ClCompile Include="..\src\network\core\tcp_game.cpp" />
@@ -1223,6 +1222,8 @@
     <ClCompile Include="..\src\network\core\tcp_http.cpp" />
     <ClInclude Include="..\src\network\core\tcp_http.h" />
     <ClInclude Include="..\src\network\core\tcp_listen.h" />
+    <ClCompile Include="..\src\network\core\tcp_packet.cpp" />
+    <ClInclude Include="..\src\network\core\tcp_packet.h" />
     <ClCompile Include="..\src\network\core\udp.cpp" />
     <ClInclude Include="..\src\network\core\udp.h" />
     <ClInclude Include="..\src\pathfinder\follow_track.hpp" />
