+++ projects/openttd_vs80.vcproj
@@ -4211,14 +4211,6 @@
 				>
 			</File>
 			<File
-				RelativePath=".\..\src\network\core\tcp.cpp"
-				>
-			</File>
-			<File
-				RelativePath=".\..\src\network\core\tcp.h"
-				>
-			</File>
-			<File
 				RelativePath=".\..\src\network\core\tcp_admin.cpp"
 				>
 			</File>
@@ -4227,7 +4219,11 @@
 				>
 			</File>
 			<File
-				RelativePath=".\..\src\network\core\tcp_connect.cpp"
+				RelativePath=".\..\src\network\core\tcp_connecter.cpp"
+				>
+			</File>
+			<File
+				RelativePath=".\..\src\network\core\tcp_connecter.h"
 				>
 			</File>
 			<File
@@ -4256,6 +4252,14 @@
 			</File>
 			<File
 				RelativePath=".\..\src\network\core\tcp_listen.h"
+				>
+			</File>
+			<File
+				RelativePath=".\..\src\network\core\tcp_packet.cpp"
+				>
+			</File>
+			<File
+				RelativePath=".\..\src\network\core\tcp_packet.h"
 				>
 			</File>
 			<File
