+++ src/network/core/tcp_connecter.cpp
@@ -16,7 +16,7 @@
 #include "../../stdafx.h"
 #include "../../thread/thread.h"
 
-#include "tcp.h"
+#include "tcp_connecter.h"
 
 #include "../../safeguards.h"
